chore: move scripts to bk_migration
